% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{default-params}
\alias{default-params}
\title{Default parameters inherited by other documentation}
\arguments{
\item{a, b}{The input strings}

\item{useNames}{Boolean. Use input vectors as row and column names?}

\item{...}{Additional arguments to be passed to \code{\link[stringdist:stringdist]{stringdist::stringdist()}} or
\code{\link[stringdist:stringsim]{stringdist::stringsim()}}.}
}
\description{
Default parameters inherited by other documentation
}
\seealso{
\code{\link[stringdist:stringdist]{stringdist::stringdist()}}, \code{\link[stringdist:stringsim]{stringdist::stringsim()}} for details on the underlying
functions and the additional options available.
}
