% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotel_rooms.R
\docType{data}
\name{hotel_rooms}
\alias{hotel_rooms}
\title{Hotel room listings}
\format{
A data frame with the following columns:
\describe{
\item{expedia}{The title of the room's listing on expedia.com}
\item{booking}{The title of the room's listing on booking.com}
}
}
\source{
Based on a dataset compiled by \href{https://github.com/susanli2016}{Susan Li}.
}
\usage{
hotel_rooms
}
\description{
The dataset contains 85 descriptions of the same hotel rooms from
Expedia and Booking.com.
}
\keyword{internal}
