% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{lev_distance}
\alias{lev_distance}
\title{Edit distance between strings}
\usage{
lev_distance(a, b, useNames = TRUE, ...)
}
\arguments{
\item{a}{The input strings}

\item{b}{The input strings}

\item{useNames}{Boolean. Use input vectors as row and column names?}

\item{...}{Additional arguments to be passed to \code{\link[stringdist:stringdist]{stringdist::stringdist()}} or
\code{\link[stringdist:stringsim]{stringdist::stringsim()}}.}
}
\value{
A numeric scalar, vector or matrix depending on the length of the inputs. See "Details".
}
\description{
Refer to \href{https://en.wikipedia.org/wiki/Levenshtein_distance}{Levenshtein distance} on
Wikipedia for more details.
}
