% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{lev_do_vectorised}
\alias{lev_do_vectorised}
\title{Calculate over vectorised input}
\usage{
lev_do_vectorised(a, b, FUN, ...)
}
\arguments{
\item{a, b}{Character vectors.}

\item{FUN}{Function to be vectorised.}

\item{...}{Other arguments to be passed to \code{FUN}.}
}
\value{
A matrix.
}
\description{
This is a utility function and is not intend to be called directly. Some operations are tricky
to vecotorise natively without ugly code, hence the need to break this out into a separate
function.
}
