% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{lev_pairwise_result}
\alias{lev_pairwise_result}
\title{Convert a matrix into a pairwise result}
\usage{
lev_pairwise_result(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A scalar or vector - see "Details".
}
\description{
Convert a matrix into a pairwise result
}
\section{Details}{

Given an input matrix, return only the diagonal. If either \code{\link[=nrow]{nrow()}} or \code{\link[=ncol]{ncol()}} are 1, return a
vector. This is a utility function and should not be called directly. Effectively this tries to
return the diagonal of a matrix and throws errors if various assumptions do not hold. The three
scenarios considered are:
\enumerate{
\item The matrix is 1 x 1: Easy, just convert it to a length 1 vector and return.
\item The matrix is 1 x n or n x 1. Convert the single row or column to a vector and return. The
elements of the vector will be named using whichever of the input vectors has length > 1.
\item The matrix is n x n. If the number of rows and columns are the same, return the diagonal as an
(unnamed) vector. If the number of rows and columns are different an error will be thrown.
}
}

