% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{lev_ratio}
\alias{lev_ratio}
\title{String similarity ratio}
\usage{
lev_ratio(a, b, pairwise = TRUE, useNames = TRUE, ...)
}
\arguments{
\item{a, b}{The input strings}

\item{pairwise}{Boolean. If \code{TRUE}, only the pairwise distances between \code{a} and \code{b} will be
computed, rather than the combinations of all elements.}

\item{useNames}{Boolean. Use input vectors as row and column names?}

\item{...}{Additional arguments to be passed to \code{\link[stringdist:stringdist]{stringdist::stringdistmatrix()}} or
\code{\link[stringdist:stringsim]{stringdist::stringsimmatrix()}}.}
}
\value{
A numeric scalar, vector or matrix depending on the length of the inputs.
}
\description{
String similarity ratio
}
\section{Details}{

This is a thin wrapper around \code{\link[stringdist:stringsim]{stringdist::stringsimmatrix()}} and mainly exists to coerce the
output into the simplest possible format (via \code{\link[=lev_simplify_matrix]{lev_simplify_matrix()}}).

The function will return the simplest possible data structure permitted by the length of the
inputs \code{a} and \code{b}. This will be a scalar if \code{a} and \code{b} are length 1, a vector if either (but
not both) is length > 1, and a matrix otherwise.
}

\examples{
lev_ratio("Bilbo", "Frodo")

lev_ratio("Bilbo", c("Frodo", "Merry"))

lev_ratio("Bilbo", c("Frodo", "Merry"), useNames = FALSE)

lev_ratio(c("Bilbo", "Gandalf"), c("Frodo", "Merry"))
}
