% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{lev_token_set_ratio}
\alias{lev_token_set_ratio}
\title{Levenshtein ratio between strings}
\usage{
lev_token_set_ratio(a, b, preprocessor = preprocessor_default())
}
\arguments{
\item{a}{The input strings}

\item{b}{The input strings}

\item{preprocessor}{Function or \code{NULL}. If a function is passed, each input will be passed
through it before analysis starts. Typical use cases are removing punctuation and standardising
case. See \code{vignette("preprocessors", package = "levitate")} for more details.}
}
\value{
A numeric scalar, vector or matrix depending on the length of the inputs. See "Details".
}
\description{
Levenshtein ratio between strings
}
