% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{levenshtein-default-params}
\alias{levenshtein-default-params}
\title{Default parameters inherited by other documentation}
\arguments{
\item{a, b}{The input strings}

\item{preprocessor}{Function or \code{NULL}. If a function is passed, each input will be passed
through it before analysis starts. Typical use cases are removing punctuation and standardising
case. See \code{vignette("preprocessors", package = "levitate")} for more details.}

\item{pairwise}{Boolean. If \code{TRUE} the function will try and compare the strings pairwise (i.e.
\code{a[1]} with \code{b[1]}, \code{a[2]} with \code{b[2]}, etc.) and return a vector of the results rather than
comparing every input with every output. For this to work, \code{a} \code{b} must be the same length or
one of them must be length 1.}
}
\description{
Default parameters inherited by other documentation
}
