% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{str_char_sum}
\alias{str_char_sum}
\title{Sum \code{nchar()}}
\usage{
str_char_sum(a, b)
}
\arguments{
\item{a, b}{The input strings. Vectorised input is accepted.}
}
\value{
The result of \code{nchar(a) + nchar(b)} for all combinations of the elements of \code{a} and \code{b}.
If \code{a} and \code{b} are length 1 the result is a scalar (i.e. a length 1 vector) otherwise a matrix.
}
\description{
Vectorised Levenshtein ratio calculations require us to add the number of characters in two
strings together. This function is vectorised over \code{a} and \code{b}. Unless both inputs are scalar the
result will be a matrix to simplify calculations in \code{\link[=lev_ratio]{lev_ratio()}}.
}
\examples{
\dontrun{
str_char_sum("cat", "mouse")
## [1] 8

str_char_sum(c("cat", "cats"), "mouse")
##      mouse
## cat      8
## cats     9

str_char_sum(c("cat", "cats"), c("mouse", "mice"))
##      mouse mice
## cat      8    7
## cats     9    8
}
}
