% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{str_token_sort}
\alias{str_token_sort}
\title{Tokenise and sort a string}
\usage{
str_token_sort(x)
}
\arguments{
\item{x}{The input string. Non-character inputs will be coerced with \code{\link[=as.character]{as.character()}}.}
}
\value{
A character vector the same length as \code{x} containing the sorted tokens.
}
\description{
Given an input string, tokenise it (using \code{\link[=str_tokenise]{str_tokenise()}}), sort the tokens alphabetically and
return the result as a single space-separated string.
}
\examples{
\dontrun{
str_token_sort("R is great fun")
#> [1] "fun great is R"

str_token_sort(c("R is great fun", "string manipulation is less so"))
#> [1] "fun great is R"                 "is less manipulation so string"
}
}
