% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{str_tokenise}
\alias{str_tokenise}
\title{Tokenise a string}
\usage{
str_tokenise(x, split = "[^[:alnum:]]+")
}
\arguments{
\item{x}{The input string.}

\item{split}{The regular expression to split on. See \code{\link[=strsplit]{strsplit()}}.}
}
\value{
A list containing one character vector for each element of \code{x}.
}
\description{
Splits an input string into tokens. A wrapper for \code{\link[=strsplit]{strsplit()}} with a predefined split regex of
\verb{[^[:alnum:]]}, i.e. anything except \verb{[a-zA-Z0-9]}.
}
\examples{
\dontrun{
str_tokenise("this is nice isn't it")
#> [[1]]
#> [1] "this" "is"   "nice" "isn"  "t"    "it"

str_tokenise(c("hello R", "goodbye R"))

#> [[1]]
#> [1] "hello" "R"

#> [[2]]
#> [1] "goodbye" "R"
}
}
