% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lev-distance.R
\name{lev_ratio}
\alias{lev_ratio}
\title{Levenshtein ratio between strings}
\usage{
lev_ratio(a, b, useNames = TRUE, ...)
}
\arguments{
\item{a}{The input strings}

\item{b}{The input strings}

\item{useNames}{Boolean. Use input vectors as row and column names?}

\item{...}{Additional arguments to be passed to \code{\link[stringdist:stringdist]{stringdist::stringdist()}} or
\code{\link[stringdist:stringsim]{stringdist::stringsim()}}.}
}
\value{
A numeric scalar, vector or matrix depending on the length of the inputs. See "Details".
}
\description{
Levenshtein ratio between strings
}
